Refactor container name and extracting duplicate logic to reusable component
