+++ src/v3/src/components/Form/Layout.tsx
@@ -17,13 +17,12 @@ import {
   AccordionLayout,
   StepperLayout,
   UISchemaElement,
-  UISchemaElementComponent,
   UISchemaLayout,
   UISchemaLayoutType,
 } from '../../types';
-import { getElementKey, isDevelopmentEnvironment, isTestEnvironment } from '../../util';
+import { getElementKey } from '../../util';
 import Accordion from './Accordion';
-import renderers from './renderers';
+import ElementContainer from './ElementContainer';
 import Stepper from './Stepper';
 
 const Layout: FunctionComponent<{ uischema: UISchemaLayout }> = ({ uischema }) => {
@@ -55,27 +54,11 @@ const Layout: FunctionComponent<{ uischema: UISchemaLayout }> = ({ uischema }) =
             return <Layout uischema={element as UISchemaLayout} />;
           }
 
-          const renderer = renderers.find((r) => r.tester(element));
-          if (!renderer) {
-            // TODO: for now do not render for unmatch render object
-            // remove unnecessary uischema in future refactor and throw error
-            // throw new Error(`Failed to find render component for uischema: ${JSON.stringify(element)}`);
-            if (isDevelopmentEnvironment() || isTestEnvironment()) {
-              console.warn(`Failed to find render component for uischema: ${JSON.stringify(element)}`);
-            }
-            return null;
-          }
-
-          const uischemaElement = (element as UISchemaElement);
-          const Component = renderer.renderer as UISchemaElementComponent;
           return (
-            <Box
+            <ElementContainer
               key={elementKey}
-              // eslint-disable-next-line react/jsx-props-no-spreading
-              {...(!(uischemaElement).noMargin && { marginBottom: 4 })}
-            >
-              <Component uischema={uischemaElement} />
-            </Box>
+              element={element as UISchemaElement}
+            />
           );
         })
       }
