+++ src/v3/src/components/Form/Stepper.tsx
@@ -15,7 +15,7 @@ import { useState } from 'preact/hooks';
 
 import { StepperContext } from '../../contexts';
 import { StepperLayout } from '../../types';
-import Layout from './SimpleLayout';
+import LayoutContainer from './LayoutContainer';
 
 type StepperProps = {
   uischema: StepperLayout;
@@ -38,7 +38,7 @@ const Stepper: FunctionComponent<StepperProps> = ({ uischema }) => {
     // Scope setStepIndex by only providing it to its own children components
     // Nested stepper can be supported with this pattern
     <StepperContext.Provider value={{ setStepIndex, stepIndex }}>
-      <Layout uischema={elements[stepIndex]} />
+      <LayoutContainer uischema={elements[stepIndex]} />
     </StepperContext.Provider>
   );
 };
